Allow getting connected IPs
Some devices don't publish correctly their hostname via DHCP, so we
cannot rely on that. IP addresses are not very reliable, but if we
assume that the user can reserve IPs for devices, we can use them as
device identifiers.

I wonder if it's possible to get the MAC address from the OnHub
diagnostic data. That would be much better.